% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.spiv.R
\name{plot.spiv}
\alias{plot.spiv}
\title{Plot for SP-IV models}
\usage{
\method{plot}{spiv}(x, type = c("outcome", "endogenous"), idx = "all", ...)
}
\arguments{
\item{x}{A \code{spiv} object.}

\item{type}{The IRF to plot. \code{"outcome"} gives the IRF of the outcome variable $y$ and \code{"endogenous"} gives the IRFs of the endogenous variables.}

\item{idx}{If \verb{type = c"outcome"}, \code{idx} must be the index of the instrument you want the IRF with respect to or \code{all} if you want a grid with every plot. If \code{type = "endogenous"}, \code{idx} must be a column vector : \verb{c(endogenous variable index, instrument index)} or \code{all}.}

\item{...}{Additional arguments.}
}
\value{
A ggplot of the IRF and HAC confidence intervals.
}
\description{
Draws a plot for \code{\link{spiv}} class models.
}
