% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiv.R
\name{spiv}
\alias{spiv}
\title{System Projections with Instrumental Variables}
\usage{
spiv(
  y,
  Y,
  X,
  Z,
  H,
  alpha = 0.05,
  wiv = c("AR", "KLM"),
  xi = 0.1,
  grid = c(lower = -1, upper = 1, length = 100),
  cores = 1
)
}
\arguments{
\item{y}{A T x 1 matrix representing the outcome variable.}

\item{Y}{A T x K matrix representing the endogenous variables.}

\item{X}{A Nx x T matrix representing the control variables.}

\item{Z}{A Nz x T matrix representing the instrumental variables.}

\item{H}{An integer specifying the maximum forecast horizon.}

\item{alpha}{A numeric value for the significance level (default is \code{0.05}).}

\item{wiv}{A character string specifying the weak instrument test to use: \code{"AR"} for Anderson-Rubin or \code{"KLM"} for Kleibergen.}

\item{xi}{A numeric value specifying the bias tolerance for the weak instrument test (default is \code{0.10}) : see section 4.3.2 of the Dallas Fed WP.}

\item{grid}{A named numeric vector (lower, upper, length) defining the bounds and resolution of the robust grid search (default is \code{c(-1, 1, 100)}).}

\item{cores}{An integer specifying the number of CPU cores to use for parallel processing during the grid search (default is \code{1}).}
}
\value{
An S3 object of class \code{spiv} containing point estimates, standard errors, weak instrument diagnostics, robust bounds, impulse response matrices and arguments used.
}
\description{
Estimates structural parameters and dynamic impulse responses using the SP-IV methodology (D. J. Lewis and K. Mertens, 2024, available at \url{https://karelmertens.com/wp-content/uploads/2024/07/dynamic_identification_using_system_projections_and_instrumental_variables_main_7_1_24.pdf}). The paper used to source the equations is the Federal Reserve Bank of Dallas working paper from March 2022 (\doi{https://doi.org/10.24149/wp2204}). The implementation uses Local Projections (section 4.1 of the Dallas Fed WP). Both AR and KLM tests are available but KLM can be very slow for large grids.
}
\examples{
\dontrun{
# Assuming y, Y, X, and Z are pre-loaded matrices
results <- spiv(y = y, Y = Y, X = X, Z = Z, H = 10, wiv = "AR", cores = 4)
summary(results)
plot(results)
}

}
