% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.spiv.R
\name{summary.spiv}
\alias{summary.spiv}
\title{Summary for SP-IV models}
\usage{
\method{summary}{spiv}(object, ...)
}
\arguments{
\item{object}{A \code{spiv} object.}
}
\value{
Prints a formatted text summary of the estimates and diagnostics.
}
\description{
Extracts a summary for \code{\link{spiv}} class models.
}
